integracion de sweetAlert2

+++ src/app/components/products/products.component.ts
@@ -4,11 +4,7 @@ import { ArticulosService } from '../../servicios/articulos.service';
 import { Articulo } from 'src/app/interfaces/articulo.interface';
 import { ToastrService } from 'ngx-toastr';
 import { NgbDate } from '@ng-bootstrap/ng-bootstrap';
-
-class ImageSnippet {
-  constructor(public src: string, public file: File) { }
-}
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-products',
@@ -27,21 +23,25 @@ export class ProductsComponent implements OnInit {
     id: 0,
     codigo: "",
     nombre: "",
-    stock: 0,
+    stock: null,
     descripcion: "",
     imagen: "",
     vencimiento: null,
     idPaquete: "",
-    idClasificacion: ""
+    idClasificacion: "",
+    idProveedor: "",
+    precioCompra: null,
+    precioVenta: null,
   };
+  title = 'sweetAlert'
   titleModal: string;
   dtItems: any[] = [];
   paquetes: any[] = [];
   clasificacion: any[] = [];
-  // loadAPI: Promise<any>;
-  selectedFile: ImageSnippet;
-  formDataFile: FormData;
+  proveedores: any[] = [];
   model: any;
+  fileToUpload: File = null;
+  // loadAPI: Promise<any>;
 
   constructor(private _contentHeaderService: ContentHeaderService,
     private _articuloService: ArticulosService,
@@ -102,6 +102,13 @@ export class ProductsComponent implements OnInit {
         this.clasificacion.push(data[key$]);
       }
     });
+
+    // proveedor
+    this._articuloService.getProveedor().subscribe(data => {
+      for (let key$ in data) {
+        this.proveedores.push(data[key$]);
+      }
+    });
   }
 
   formatDate(date: any) {
@@ -119,12 +126,15 @@ export class ProductsComponent implements OnInit {
       id: 0,
       codigo: "",
       nombre: "",
-      stock: 0,
+      stock: null,
       descripcion: "",
       imagen: "",
       vencimiento: null,
       idPaquete: "",
-      idClasificacion: ""
+      idClasificacion: "",
+      idProveedor: "",
+      precioCompra: null,
+      precioVenta: null,
     };
   }
 
@@ -146,32 +156,43 @@ export class ProductsComponent implements OnInit {
       imagen: art.Imagen,
       vencimiento: new Date(currentDate.join('/')),
       idPaquete: art.IdPaquete,
-      idClasificacion: art.IdClasificacion
+      idClasificacion: art.IdClasificacion,
+      idProveedor: art.IdProveedor,
+      precioCompra: art.PrecioCompra,
+      precioVenta: art.PrecioVenta
     };
   }
 
   deleteItem(art: any) {
-    this._articuloService.eliminarArticulo(art.Id).subscribe(status => {
-      if (status) {
-        this.loadDataProducts();
+    Swal.fire({
+      title: `Desea eliminar ${art.Nombre}?`,
+      text: "Esta accion no puede revertirse!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this._articuloService.eliminarArticulo(art.Id).subscribe(status => {
+          if (status == 'success') {
+            Swal.fire(
+              'Eliminado!',
+              'Registro eliminado de manera exitosa.',
+              'success'
+            )
+            this.loadDataProducts();
+          }
+        });
+      } else {
+
       }
-    });
+    })
   }
 
-  processFile(img: any) {
-    // debugger;
-    const file: File = img.files[0];
-    const reader = new FileReader();
-
-    reader.addEventListener('load', (event: any) => {
-      this.selectedFile = new ImageSnippet(event.target.result, file);
-      const formData = new FormData();
-      const image: File = this.selectedFile.file;
-      formData.append('image', image);
-      this.item.formData = formData;
-      this.formDataFile = formData;
-    });
-
+  processFile(files: FileList) {
+      this.fileToUpload = files.item(0);
   }
 
   guardar() {
@@ -186,7 +207,7 @@ export class ProductsComponent implements OnInit {
       this.item.vencimiento = new Date(selectedDate.join('/'));
 
       // ejecuta el servicio para guardar la informacion en el servidor
-      this._articuloService.nuevoArticulo(this.item).subscribe(data => {
+      this._articuloService.nuevoArticulo(this.item, this.fileToUpload).subscribe(data => {
         if (data == 'success') {
           var modal = document.getElementById("closeBtn") as any;
           modal.click();
@@ -214,6 +235,5 @@ export class ProductsComponent implements OnInit {
         }
       });
     }
-
   }
 }
